refactor(pdf-overlay): extract prefetch and scheduler coordination
